# api/codesurgeon.openapi.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "codesurgeon",
		"version": "1.0.0"
	},
	"paths": {
		"/codesurgeon.GptService/GetOpenAPI": {
			"post": {
				"summary": "GetOpenAPI",
				"operationId": "codesurgeon.GptService.GetOpenAPI",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/codesurgeon.GetOpenAPIRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "A successful response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/codesurgeon.GetOpenAPIResponse"
								}
							}
						}
					}
				}
			}
		},
		"/codesurgeon.GptService/NewChat": {
			"post": {
				"summary": "NewChat",
				"operationId": "codesurgeon.GptService.NewChat",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/codesurgeon.NewChatRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "A successful response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/codesurgeon.NewChatResponse"
								}
							}
						}
					}
				}
			}
		},
		"/codesurgeon.GptService/GetChat": {
			"post": {
				"summary": "GetChat",
				"operationId": "codesurgeon.GptService.GetChat",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/codesurgeon.GetChatRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "A successful response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/codesurgeon.GetChatResponse"
								}
							}
						}
					}
				}
			}
		},
		"/codesurgeon.GptService/SendMessage": {
			"post": {
				"summary": "SendMessage",
				"operationId": "codesurgeon.GptService.SendMessage",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/codesurgeon.SendMessageRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "A successful response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/codesurgeon.SendMessageResponse"
								}
							}
						}
					}
				}
			}
		},
		"/codesurgeon.GptService/ReceiveSlackMessage": {
			"post": {
				"summary": "ReceiveSlackMessage",
				"operationId": "codesurgeon.GptService.ReceiveSlackMessage",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/codesurgeon.ReceiveSlackMessageRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "A successful response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/codesurgeon.ReceiveSlackMessageResponse"
								}
							}
						}
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
			"codesurgeon.GetChatRequest": {
				"type": "object",
				"properties": {
					"chat_id": {
						"type": "string"
					}
				}
			},
			"codesurgeon.GetChatResponse": {
				"type": "object",
				"properties": {
					"chat": {
						"$ref": "#/components/schemas/codesurgeon.Chat"
					}
				}
			},
			"codesurgeon.NewChatRequest": {
				"type": "object",
				"properties": {
					"external_id": {
						"type": "string"
					}
				}
			},
			"codesurgeon.NewChatResponse": {
				"type": "object",
				"properties": {
					"chat": {
						"$ref": "#/components/schemas/codesurgeon.Chat"
					}
				}
			},
			"codesurgeon.Chat": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"current_mode": {
						"$ref": "#/components/schemas/codesurgeon.Mode"
					},
					"mode_state": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/codesurgeon.ModeState"
						}
					},
					"messages": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/codesurgeon.Message"
						}
					}
				}
			},
			"codesurgeon.ModeState": {
				"type": "object",
				"properties": {
					"mode_name": {
						"type": "string"
					},
					"key": {
						"type": "string"
					},
					"value": {
						"type": "string"
					}
				}
			},
			"codesurgeon.SendMessageRequest": {
				"type": "object",
				"properties": {
					"chat_id": {
						"type": "string"
					},
					"message": {
						"$ref": "#/components/schemas/codesurgeon.Message"
					}
				}
			},
			"codesurgeon.SendMessageResponse": {
				"type": "object",
				"properties": {
					"chat_id": {
						"type": "string"
					},
					"command": {
						"$ref": "#/components/schemas/codesurgeon.Command"
					},
					"message": {
						"$ref": "#/components/schemas/codesurgeon.Message"
					},
					"mode": {
						"$ref": "#/components/schemas/codesurgeon.Mode"
					}
				}
			},
			"codesurgeon.Mode": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string"
					}
				}
			},
			"codesurgeon.Command": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string"
					}
				}
			},
			"codesurgeon.Message": {
				"type": "object",
				"properties": {
					"text": {
						"type": "string"
					},
					"form": {
						"$ref": "#/components/schemas/codesurgeon.FormMessage"
					},
					"sender": {
						"type": "string"
					}
				}
			},
			"codesurgeon.FormMessage": {
				"type": "object",
				"properties": {
					"questions": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/codesurgeon.QuestionAnswer"
						}
					}
				}
			},
			"codesurgeon.QuestionAnswer": {
				"type": "object",
				"properties": {
					"question": {
						"type": "string"
					},
					"answer": {
						"type": "string"
					}
				}
			},
			"codesurgeon.GetOpenAPIRequest": {
				"type": "object",
				"properties": {
				}
			},
			"codesurgeon.GetOpenAPIResponse": {
				"type": "object",
				"properties": {
					"openapi": {
						"type": "string"
					}
				}
			},
			"codesurgeon.ReceiveSlackMessageRequest": {
				"type": "object",
				"properties": {
					"token": {
						"type": "string"
					},
					"challenge": {
						"type": "string"
					},
					"type": {
						"type": "string"
					}
				}
			},
			"codesurgeon.ReceiveSlackMessageResponse": {
				"type": "object",
				"properties": {
					"challenge": {
						"type": "string"
					}
				}
			}
		}
	}
}